chore: moving utility_tree to prebuilt, getting this into our serialization / testing of that framework
